pyproject.toml
[project]
name = "iskra-etl"
version = "0.1.0"
description = "Iskra 项目的 ETL 部分"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "llama-index-core>=0.14.21",
    "numpy>=2.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.0.0",
    "torch>=2.11.0",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

# Win：`torch` 轮子来自 PyTorch CUDA 索引；与驱动不符时改用 cu128/cu130 并重跑 uv lock。
[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
